agents_templates/pydantic_ai/requirements.txt
pydantic_ai
azure-ai-inference 
aiohttp
mcp
fastapi
uvicorn
jsonschema
openai